extracted GuiSelectDimension
